style: move directory store code
