+++ src/core/presentation/hooks/stores/tripStore.ts
